Route valorlegacies.xyz to client portal instead of leads homepage
The portal domain (valorlegacies.xyz) was serving the same leads
homepage as the marketing domains. Now:

- Middleware redirects all non-portal routes on the portal domain
  (/, /blog, /about, etc.) to /portal
- New welcome page at /portal with centered "Welcome, Clients"
  heading, login form, feature cards, and quick links
- Moved the leads marketplace from /portal to /portal/marketplace
- Navbar shows portal-specific navigation on the portal domain
  (Home, Dashboard, Leads Marketplace, Terms, Privacy)
- Admin button always visible in the portal domain header
- Portal login redirects to /portal/dashboard after auth
