+++ digital-cathedral/app/components/navbar.tsx
@@ -39,6 +39,14 @@ const NAV_LINKS = [
   { href: "/terms", label: "Terms of Service" },
 ];
 
+const PORTAL_NAV_LINKS = [
+  { href: "/portal", label: "Home" },
+  { href: "/portal/dashboard", label: "Dashboard" },
+  { href: "/portal/marketplace", label: "Leads Marketplace" },
+  { href: "/portal/terms", label: "Terms of Service" },
+  { href: "/portal/privacy", label: "Privacy Policy" },
+];
+
 export function Navbar() {
   const isAdmin = useIsAdmin();
   const isPortalDomain = useIsPortalDomain();
@@ -149,7 +157,7 @@ export function Navbar() {
               aria-label="Main navigation menu"
               className="absolute left-0 top-full mt-fib-3 w-56 rounded-[13px] py-fib-5 z-50 cathedral-surface"
             >
-              {NAV_LINKS.map((link) => (
+              {(isPortalDomain ? PORTAL_NAV_LINKS : NAV_LINKS).map((link) => (
                 <Link
                   key={link.href}
                   href={link.href}
@@ -178,16 +186,16 @@ export function Navbar() {
             <span className="text-sm text-[var(--text-primary)] hidden sm:inline">
               {session.user.name?.split(" ")[0]}
             </span>
-            {Boolean((session.user as Record<string, unknown>).isAdmin) && (
+            {(Boolean((session.user as Record<string, unknown>).isAdmin) || isPortalDomain) && (
               <Link
                 href="/admin"
-                className="text-xs text-teal-cathedral hover:text-teal-cathedral/80 transition-colors"
+                className="flex items-center gap-fib-5 px-fib-13 py-fib-5 text-xs font-medium rounded-fib border border-[var(--teal)]/30 text-[var(--teal)] hover:border-[var(--teal)] transition-all"
               >
                 Admin
               </Link>
             )}
             <button
-              onClick={() => signOut({ callbackUrl: "/" })}
+              onClick={() => signOut({ callbackUrl: isPortalDomain ? "/portal" : "/" })}
               className="text-xs text-[var(--text-muted)] hover:text-[var(--teal)] transition-colors"
             >
               Sign Out
@@ -197,7 +205,7 @@ export function Navbar() {
           <div className="flex items-center gap-fib-8">
             <Link
               href="/admin/login"
-              className="flex items-center gap-fib-5 px-fib-13 py-fib-5 text-xs font-medium rounded-fib border border-[var(--teal)] text-[var(--teal)] hover:bg-[var(--teal)] hover:text-[var(--bg-primary)] transition-all"
+              className="flex items-center gap-fib-5 px-fib-13 py-fib-5 text-xs font-medium rounded-fib border border-[var(--teal)]/30 text-[var(--teal)] hover:border-[var(--teal)] transition-all"
             >
               <svg
                 width="14"
@@ -211,11 +219,11 @@ export function Navbar() {
               >
                 <path d="M9 12.75L11.25 15 15 9.75m-3-7.036A11.959 11.959 0 013.598 6 11.99 11.99 0 003 9.749c0 5.592 3.824 10.29 9 11.623 5.176-1.332 9-6.03 9-11.622 0-1.31-.21-2.571-.598-3.751h-.152c-3.196 0-6.1-1.248-8.25-3.285z" />
               </svg>
-              Admin Login
+              Admin
             </Link>
             <Link
               href="/portal/login"
-              className="flex items-center gap-fib-5 px-fib-13 py-fib-5 text-xs font-medium rounded-fib border border-[var(--teal)] text-[var(--teal)] hover:bg-[var(--teal)] hover:text-[var(--bg-primary)] transition-all"
+              className="flex items-center gap-fib-5 px-fib-13 py-fib-5 text-xs font-medium rounded-fib bg-teal-cathedral text-white hover:bg-teal-cathedral/90 transition-all"
             >
               <svg
                 width="14"
@@ -227,7 +235,7 @@ export function Navbar() {
                 className="shrink-0"
                 aria-hidden="true"
               >
-                <path d="M20.25 14.15v4.25c0 1.094-.787 2.036-1.872 2.18-2.087.277-4.216.42-6.378.42s-4.291-.143-6.378-.42c-1.085-.144-1.872-1.086-1.872-2.18v-4.25m16.5 0a2.18 2.18 0 00.75-1.661V8.706c0-1.081-.768-2.015-1.837-2.175a48.114 48.114 0 00-3.413-.387m4.5 8.006c-.194.165-.42.295-.673.38A23.978 23.978 0 0112 15.75c-2.648 0-5.195-.429-7.577-1.22a2.016 2.016 0 01-.673-.38m0 0A2.18 2.18 0 013 12.489V8.706c0-1.081.768-2.015 1.837-2.175a48.111 48.111 0 013.413-.387m7.5 0V5.25A2.25 2.25 0 0013.5 3h-3a2.25 2.25 0 00-2.25 2.25v.894m7.5 0a48.667 48.667 0 00-7.5 0M12 12.75h.008v.008H12v-.008z" />
+                <path d="M15.75 9V5.25A2.25 2.25 0 0013.5 3h-6a2.25 2.25 0 00-2.25 2.25v13.5A2.25 2.25 0 007.5 21h6a2.25 2.25 0 002.25-2.25V15m3 0l3-3m0 0l-3-3m3 3H9" />
               </svg>
               Client Login
             </Link>
